+++ src/plugin/names.ml
@@ -8,8 +8,7 @@ let is_reserved = function
   | "land" | "lazy" | "let" | "lor" | "lsl" | "lsr" | "lxor" | "match" | "method"
   | "mod" | "module" | "mutable" | "new" | "nonrec" | "object" | "of" | "open" | "or"
   | "private" | "rec" | "sig" | "struct" | "then" | "to" | "true" | "try" | "type"
-  | "val" | "virtual" | "when" | "while" | "with" ->
-    true
+  | "val" | "virtual" | "when" | "while" | "with" -> true
   | _ -> false
 
 let module_name name =